logMigration/requirements.txt
pymongo==3.9.0
mysql-connector==2.1.7
pytz==2019.3
